pyproject.toml
[tool.poetry]
name = "q8s"
version = "0.1.0"
description = "Emulated Kubernetes Cluster"
authors = ["Vincent Hasse <v.hasse@stud.uni-goettingen.de>"]
license = "MIT"
readme = "README.md"
packages = [{include = "q8s", from = "src"}]

[tool.poetry.scripts]
q8s = "q8s.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
click = "^8.1.7"
openstacksdk = "^3.1.0"
keystoneauth1 = "^5.6.0"
paramiko = "^3.4.0"
kubernetes = "^29.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}